requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
black==22.12.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
crispy-bootstrap5==0.6
cryptography==39.0.0
defusedxml==0.7.1
dj-database-url==1.2.0
Django==4.1.5
django-allauth==0.52.0
django-celery-results==2.4.0
django-crispy-forms==1.14.0
environs==9.5.0
idna==3.4
kombu==5.2.4
marshmallow==3.19.0
mypy-extensions==0.4.3
oauthlib==3.2.2
packaging==23.0
pathspec==0.11.0
platformdirs==2.6.2
prompt-toolkit==3.0.36
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2022.7.1
redis==4.4.2
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
django-debug-toolbar==3.8.1
whitenoise==6.4.0
gunicorn==20.1.0